fixed Visual C++ project

+++ include/vorbis/vorbisfile.h
@@ -53,7 +53,7 @@ typedef struct {
 static int _ov_header_fseek_wrap(FILE *f,ogg_int64_t off,int whence){
   if(f==NULL)return(-1);
 
-  return fseek(f,off,whence);
+  return fseek(f,long(off),whence);
 }
 
 /* These structs below (OV_CALLBACKS_DEFAULT etc) are defined here as
